+++ ocrd_models/ocrd_models/ocrd_agent.py
@@ -2,7 +2,7 @@
 API to ``mets:agent``
 """
 #  import os
-from .constants import NAMESPACES as NS, TAG_METS_AGENT, TAG_METS_NAME
+from .constants import NAMESPACES as NS, TAG_METS_AGENT, TAG_METS_NAME, TAG_METS_NOTE
 from .ocrd_xml_base import ET
 
 class OcrdAgent():
@@ -20,7 +20,8 @@ class OcrdAgent():
     #      #  version = name_parts[1][1:]     # v0.0.1 => 0.0.1
     #      return OcrdAgent(el, name, role, _type, otherrole)
 
-    def __init__(self, el=None, name=None, _type=None, othertype=None, role=None, otherrole=None):
+    def __init__(self, el=None, name=None, _type=None, othertype=None, role=None, otherrole=None,
+                 notes=None):
         """
         Args:
             el (LxmlElement):
@@ -29,6 +30,7 @@ class OcrdAgent():
             othertype (string):
             role (string):
             otherrole (string):
+            notes (dict):
         """
         if el is None:
             el = ET.Element(TAG_METS_AGENT)
@@ -38,6 +40,7 @@ class OcrdAgent():
         self.othertype = othertype
         self.role = role
         self.otherrole = otherrole
+        self.notes = notes
 
     def __str__(self):
         """
@@ -105,7 +108,7 @@ class OcrdAgent():
     @otherrole.setter
     def otherrole(self, otherrole):
         """
-        Get the ``OTHERROLE`` attribute value.
+        Set the ``OTHERROLE`` attribute value.
         """
         if otherrole is not None:
             self._el.set('ROLE', 'OTHER')
@@ -116,17 +119,44 @@ class OcrdAgent():
         """
         Get the ``mets:name`` element value.
         """
-        el_name = self._el.find('mets:name', NS)
+        el_name = self._el.find(TAG_METS_NAME)
         if el_name is not None:
             return el_name.text
 
     @name.setter
     def name(self, name):
         """
-        Get the ``mets:name`` element value.
+        Set the ``mets:name`` element value.
         """
         if name is not None:
-            el_name = self._el.find('mets:name', NS)
+            el_name = self._el.find(TAG_METS_NAME)
             if el_name is None:
                 el_name = ET.SubElement(self._el, TAG_METS_NAME)
             el_name.text = name
+
+    @property
+    def notes(self):
+        """
+        Get the ``mets:note`` element values (as tuples of attributes and text).
+        """
+        el_notes = self._el.findall(TAG_METS_NOTE)
+        if el_notes is not None:
+            return [(note.attrib, note.text)
+                    for note in el_notes]
+
+    @notes.setter
+    def notes(self, notes):
+        """
+        Set the ``mets:note`` element values.
+        """
+        el_notes = self._el.findall(TAG_METS_NOTE)
+        if el_notes:
+            for el_note in el_notes:
+                self._el.remove(el_note)
+        if notes:
+            for note in notes:
+                el_note = ET.SubElement(self._el, TAG_METS_NOTE, nsmap={'ocrd': NS['ocrd']})
+                attrib, text = note
+                el_note.text = text
+                for name, value in attrib.items():
+                    el_note.set('{%s}' % NS["ocrd"] + name, value)