+++ ocrd_models/ocrd_models/ocrd_mets.py
@@ -103,7 +103,7 @@ class OcrdMets(OcrdXmlDocument):
 
     def add_agent(self, *args, **kwargs):
         """
-        Add an :py:class:`ocrd_models.ocrd_agent.OcrdAgent` to the list of agents in the `metsHdr`.
+        Add an :py:class:`ocrd_models.ocrd_agent.OcrdAgent` to the list of agents in the ``metsHdr``.
         """
         el_metsHdr = self._tree.getroot().find('.//mets:metsHdr', NS)
         if el_metsHdr is None:
@@ -132,25 +132,27 @@ class OcrdMets(OcrdXmlDocument):
     # pylint: disable=multiple-statements
     def find_files(self, ID=None, fileGrp=None, pageId=None, mimetype=None, url=None, local_only=False):
         """
-        Search `mets:file` entries in this METS document and yield results.
+        Search ``mets:file`` entries in this METS document and yield results.
 
 
-        The ``ID``, ``fileGrp``, ``url`` and ``mimetype`` parameters can each be
-        either a literal string or a regular expression if the string starts
-        with `//` (double slash). If it is a regex, the leading `//` is removed
-        and candidates are matched against the regex with `re.fullmatch`. If it is
-        a literal string, comparison is done with string equality.
+        The :py:attr:`ID`, :py:attr:`fileGrp`, :py:attr:`url` and :py:attr:`mimetype`
+        parameters can each be either a literal string, or a regular expression if
+        the string starts with ``//`` (double slash).
 
-        The ``pageId`` parameter supports the numeric range operator ``..``. For
+        If it is a regex, the leading ``//`` is removed and candidates are matched
+        against the regex with `re.fullmatch`. If it is a literal string, comparison
+        is done with string equality.
+
+        The :py:attr:`pageId` parameter supports the numeric range operator ``..``. For
         example, to find all files in pages ``PHYS_0001`` to ``PHYS_0003``, 
         ``PHYS_0001..PHYS_0003`` will be expanded to ``PHYS_0001,PHYS_0002,PHYS_0003``.
 
         Keyword Args:
-            ID (string) : `@ID` of the `mets:file`
-            fileGrp (string) : `@USE` of the `mets:fileGrp` to list files of
-            pageId (string) : `@ID` of the corresponding physical `mets:structMap` entry (physical page)
-            url (string) : `@xlink:href` (URL or path) of `mets:Flocat` of `mets:file`
-            mimetype (string) : `@MIMETYPE` of `mets:file`
+            ID (string) : ``@ID`` of the ``mets:file``
+            fileGrp (string) : ``@USE`` of the ``mets:fileGrp`` to list files of
+            pageId (string) : ``@ID`` of the corresponding physical ``mets:structMap`` entry (physical page)
+            url (string) : ``@xlink:href`` (URL or path) of ``mets:Flocat`` of ``mets:file``
+            mimetype (string) : ``@MIMETYPE`` of ``mets:file``
             local (boolean) : Whether to restrict results to local files in the filesystem
 
         Yields:
@@ -212,10 +214,10 @@ class OcrdMets(OcrdXmlDocument):
 
     def add_file_group(self, fileGrp):
         """
-        Add a new `mets:fileGrp`.
+        Add a new ``mets:fileGrp``.
 
         Arguments:
-            fileGrp (string): `@USE` of the new `mets:fileGrp`.
+            fileGrp (string): ``@USE`` of the new ``mets:fileGrp``.
         """
         if ',' in fileGrp:
             raise Exception('fileGrp must not contain commas')
@@ -230,7 +232,7 @@ class OcrdMets(OcrdXmlDocument):
 
     def rename_file_group(self, old, new):
         """
-        Rename a `mets:fileGrp` by changing the `@USE` from ``old`` to ``new``.
+        Rename a ``mets:fileGrp`` by changing the ``@USE`` from :py:attr:`old` to :py:attr:`new`.
         """
         el_fileGrp = self._tree.getroot().find('mets:fileSec/mets:fileGrp[@USE="%s"]' % old, NS)
         if el_fileGrp is None:
@@ -239,12 +241,12 @@ class OcrdMets(OcrdXmlDocument):
 
     def remove_file_group(self, USE, recursive=False, force=False):
         """
-        Remove a `mets:fileGrp` (fixed `@USE`) or `mets:fileGrp`s (regex `@USE`)
+        Remove a ``mets:fileGrp`` (single fixed ``@USE`` or multiple regex ``@USE``)
 
         Arguments:
-            USE (string): `@USE` of the `mets:fileGrp` to delete. Can be a regex if prefixed with `//`
-            recursive (boolean): Whether to recursively delete all `mets:file`s in the group
-            force (boolean): Do not raise an exception if `mets:fileGrp` does not exist
+            USE (string): ``@USE`` of the ``mets:fileGrp`` to delete. Can be a regex if prefixed with ``//``
+            recursive (boolean): Whether to recursively delete each ``mets:file`` in the group
+            force (boolean): Do not raise an exception if ``mets:fileGrp`` does not exist
         """
         log = getLogger('ocrd_models.ocrd_mets.remove_file_group')
         el_fileSec = self._tree.getroot().find('mets:fileSec', NS)
@@ -279,40 +281,45 @@ class OcrdMets(OcrdXmlDocument):
         Instantiate and add a new :py:class:`ocrd_models.ocrd_file.OcrdFile`.
 
         Arguments:
-            fileGrp (string): `@USE` of `mets:fileGrp` to add to
+            fileGrp (string): ``@USE`` of ``mets:fileGrp`` to add to
         Keyword Args:
-            mimetype (string): `@MIMETYPE` of the `mets:file` to use
-            url (string): `@xlink:href` (URL or path) of the `mets:file` to use
-            ID (string): `@ID` of the `mets:file` to use
-            pageId (string): `@ID` in the physical `mets:structMap` to link to
-            force (boolean): Whether to add the file even if a `mets:file` with the same `@ID` already exists.
+            mimetype (string): ``@MIMETYPE`` of the ``mets:file`` to use
+            url (string): ``@xlink:href`` (URL or path) of the ``mets:file`` to use
+            ID (string): ``@ID`` of the ``mets:file`` to use
+            pageId (string): ``@ID`` in the physical ``mets:structMap`` to link to
+            force (boolean): Whether to add the file even if a ``mets:file`` with the same ``@ID`` already exists.
             ignore (boolean): Do not look for existing files at all. Shift responsibility for preventing errors from duplicate ID to the user.
             local_filename (string):
         """
         if not ID:
-            raise Exception("Must set ID of the mets:file")
-        elif not REGEX_FILE_ID.fullmatch(ID):
-            raise Exception("Invalid syntax for mets:file/@ID %s" % ID)
+            raise ValueError("Must set ID of the mets:file")
+        if not fileGrp:
+            raise ValueError("Must set fileGrp of the mets:file")
+        if not REGEX_FILE_ID.fullmatch(ID):
+            raise ValueError("Invalid syntax for mets:file/@ID %s (not an xs:ID)" % ID)
+        if not REGEX_FILE_ID.fullmatch(fileGrp):
+            raise ValueError("Invalid syntax for mets:fileGrp/@USE %s (not an xs:ID)" % ID)
         el_fileGrp = self._tree.getroot().find(".//mets:fileGrp[@USE='%s']" % (fileGrp), NS)
         if el_fileGrp is None:
             el_fileGrp = self.add_file_group(fileGrp)
-        if ID and not ignore and next(self.find_files(ID=ID), None):
+        mets_file = next(self.find_files(ID=ID), None)
+        if mets_file and not ignore:
             if not force:
                 raise Exception("File with ID='%s' already exists" % ID)
-            mets_file = next(self.find_files(ID=ID))
+            mets_file.url = url
+            mets_file.mimetype = mimetype
+            mets_file.ID = ID
+            mets_file.pageId = pageId
+            mets_file.local_filename = local_filename
         else:
-            mets_file = OcrdFile(ET.SubElement(el_fileGrp, TAG_METS_FILE), mets=self)
-        mets_file.url = url
-        mets_file.mimetype = mimetype
-        mets_file.ID = ID
-        mets_file.pageId = pageId
-        mets_file.local_filename = local_filename
+            kwargs = {k: v for k, v in locals().items() if k in ['url', 'ID', 'mimetype', 'pageId', 'local_filename'] and v}
+            mets_file = OcrdFile(ET.SubElement(el_fileGrp, TAG_METS_FILE), mets=self, **kwargs)
 
         return mets_file
 
     def remove_file(self, *args, **kwargs):
         """
-        Delete all `ocrd:file`s matching the query. Same arguments as :py:meth:`find_files`
+        Delete each ``ocrd:file`` matching the query. Same arguments as :py:meth:`find_files`
         """
         files = list(self.find_files(*args, **kwargs))
         if files:
@@ -322,6 +329,10 @@ class OcrdMets(OcrdXmlDocument):
                 return files
             else:
                 return files[0] # for backwards-compatibility
+        if any(1 for kwarg in kwargs
+               if isinstance(kwarg, str) and kwarg.startswith(REGEX_PREFIX)):
+            # allow empty results if filter criteria involve a regex
+            return []
         raise FileNotFoundError("File not found: %s %s" % (args, kwargs))
 
     def remove_one_file(self, ID):
@@ -329,7 +340,7 @@ class OcrdMets(OcrdXmlDocument):
         Delete an existing :py:class:`ocrd_models.ocrd_file.OcrdFile`.
         
         Arguments:
-            ID (string): `@ID` of the `mets:file` to delete
+            ID (string): ``@ID`` of the ``mets:file`` to delete
             
         Returns:
             The old :py:class:`ocrd_models.ocrd_file.OcrdFile` reference.
@@ -364,7 +375,7 @@ class OcrdMets(OcrdXmlDocument):
     @property
     def physical_pages(self):
         """
-        List all page IDs (the `@ID`s of all physical `mets:structMap` `mets:div`s)
+        List all page IDs (the ``@ID`` of each physical ``mets:structMap`` ``mets:div``)
         """
         return self._tree.getroot().xpath(
             'mets:structMap[@TYPE="PHYSICAL"]/mets:div[@TYPE="physSequence"]/mets:div[@TYPE="page"]/@ID',
@@ -372,8 +383,8 @@ class OcrdMets(OcrdXmlDocument):
 
     def get_physical_pages(self, for_fileIds=None):
         """
-        List all page IDs (the `@ID`s of all physical `mets:structMap` `mets:div`s),
-        optionally for a subset of `mets:file` `@ID`s ``for_fileIds``.
+        List all page IDs (the ``@ID`` of each physical ``mets:structMap`` ``mets:div``),
+        optionally for a subset of ``mets:file`` ``@ID`` :py:attr:`for_fileIds`.
         """
         if for_fileIds is None:
             return self.physical_pages
@@ -388,15 +399,15 @@ class OcrdMets(OcrdXmlDocument):
 
     def set_physical_page_for_file(self, pageId, ocrd_file, order=None, orderlabel=None):
         """
-        Set the physical page ID (`@ID` of the physical `mets:structMap` `mets:div` entry)
-        corresponding to the `mets:file` ``ocrd_file``, creating all structures if necessary.
+        Set the physical page ID (``@ID`` of the physical ``mets:structMap`` ``mets:div`` entry)
+        corresponding to the ``mets:file`` :py:attr:`ocrd_file`, creating all structures if necessary.
         
         Arguments:
-            pageId (string): `@ID` of the physical `mets:structMap` entry to use
+            pageId (string): ``@ID`` of the physical ``mets:structMap`` entry to use
             ocrd_file (object): existing :py:class:`ocrd_models.ocrd_file.OcrdFile` object
         Keyword Args:
-            order (string): `@ORDER` to use
-            orderlabel (string): `@ORDERLABEL` to use
+            order (string): ``@ORDER`` to use
+            orderlabel (string): ``@ORDERLABEL`` to use
         """
         #  print(pageId, ocrd_file)
         # delete any page mapping for this file.ID
@@ -428,8 +439,8 @@ class OcrdMets(OcrdXmlDocument):
 
     def get_physical_page_for_file(self, ocrd_file):
         """
-        Get the physical page ID (`@ID` of the physical `mets:structMap` `mets:div` entry)
-        corresponding to the `mets:file` ``ocrd_file``.
+        Get the physical page ID (``@ID`` of the physical ``mets:structMap`` ``mets:div`` entry)
+        corresponding to the ``mets:file`` :py:attr:`ocrd_file`.
         """
         ret = self._tree.getroot().xpath(
             '/mets:mets/mets:structMap[@TYPE="PHYSICAL"]/mets:div[@TYPE="physSequence"]/mets:div[@TYPE="page"][./mets:fptr[@FILEID="%s"]]/@ID' %
@@ -439,7 +450,7 @@ class OcrdMets(OcrdXmlDocument):
 
     def remove_physical_page(self, ID):
         """
-        Delete page (physical `mets:structMap` `mets:div` entry `@ID`) ``ID``.
+        Delete page (physical ``mets:structMap`` ``mets:div`` entry ``@ID``) :py:attr:`ID`.
         """
         mets_div = self._tree.getroot().xpath(
             'mets:structMap[@TYPE="PHYSICAL"]/mets:div[@TYPE="physSequence"]/mets:div[@TYPE="page"][@ID="%s"]' % ID,
@@ -447,14 +458,30 @@ class OcrdMets(OcrdXmlDocument):
         if mets_div:
             mets_div[0].getparent().remove(mets_div[0])
 
+    def remove_physical_page_fptr(self, fileId):
+        """
+        Delete all ``mets:fptr[@FILEID = fileId]`` to ``mets:file[@ID == fileId]`` for :py:attr:`fileId` from all ``mets:div`` entries in the physical ``mets:structMap``.
+        Returns:
+            List of pageIds that mets:fptrs were deleted from
+        """
+        mets_fptrs = self._tree.getroot().xpath(
+            'mets:structMap[@TYPE="PHYSICAL"]/mets:div[@TYPE="physSequence"]/mets:div[@TYPE="page"]/mets:fptr[@FILEID="%s"]' % fileId,
+            namespaces=NS)
+        ret = []
+        for mets_fptr in mets_fptrs:
+            mets_div = mets_fptr.getparent()
+            ret.append(mets_div.get('ID'))
+            mets_div.remove(mets_fptr)
+        return ret
+
     def merge(self, other_mets, fileGrp_mapping=None, after_add_cb=None, **kwargs):
         """
         Add all files from other_mets.
 
-        Accepts the same kwargs as :py:func:find_files
+        Accepts the same kwargs as :py:func:`find_files`
 
         Keyword Args:
-            fileGrp_mapping (dict): Map ``other_mets`` fileGrp to fileGrp in this METS
+            fileGrp_mapping (dict): Map :py:attr:`other_mets` fileGrp to fileGrp in this METS
             after_add_cb (function): Callback received after file is added to the METS
         """
         if not fileGrp_mapping: