+++ ocrd_models/ocrd_page_user_methods/get_UnorderedGroupChildren.py
@@ -1,6 +1,6 @@
 def get_UnorderedGroupChildren(self):
     """
-    List all non-metadata children of an UnorderedGroup
+    List all non-metadata children of an :py:class:`UnorderedGroupType`
     """
     # TODO: should not change order
     return self.get_RegionRef() + self.get_OrderedGroup() + self.get_UnorderedGroup()