+++ ocrd_models/ocrd_page_user_methods/id.py
@@ -1,0 +1,5 @@
+@property
+def id(self):
+    if hasattr(self, 'pcGtsId'):
+        return self.pcGtsId or ''
+    return self.imageFilename